New replay only version (working with DQN)
