+++ stable_baselines3/her/__init__.py
@@ -1,4 +1,2 @@
-from stable_baselines3.common.vec_env.obs_dict_wrapper import ObsDictWrapper
 from stable_baselines3.her.goal_selection_strategy import GoalSelectionStrategy
-from stable_baselines3.her.her import HER
 from stable_baselines3.her.her_replay_buffer import HerReplayBuffer
